// src/api/generated/models/TaskPriority.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Manager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.8.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TaskPriority = {
    None: 'none',
    Important: 'important',
    Urgent: 'urgent',
    ImportantUrgent: 'important_urgent'
} as const;
export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export function instanceOfTaskPriority(value: any): boolean {
    for (const key in TaskPriority) {
        if (Object.prototype.hasOwnProperty.call(TaskPriority, key)) {
            if (TaskPriority[key as keyof typeof TaskPriority] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TaskPriorityFromJSON(json: any): TaskPriority {
    return TaskPriorityFromJSONTyped(json, false);
}

export function TaskPriorityFromJSONTyped(json: any, ignoreDiscriminator: boolean): TaskPriority {
    return json as TaskPriority;
}

export function TaskPriorityToJSON(value?: TaskPriority | null): any {
    return value as any;
}

export function TaskPriorityToJSONTyped(value: any, ignoreDiscriminator: boolean): TaskPriority {
    return value as TaskPriority;
}